Move TreeExtensions to own trait
